+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [2.11.5](https://github.com/ducktors/turborepo-remote-cache/compare/v2.11.4...v2.11.5) (2026-07-21)
+
+
+### Bug Fixes
+
+* override shell-quote ([e03a32b](https://github.com/ducktors/turborepo-remote-cache/commit/e03a32ba712f94fd6638f2e5f5b5eb12b8b2fb00))
+
+## [2.11.4](https://github.com/ducktors/turborepo-remote-cache/compare/v2.11.3...v2.11.4) (2026-07-21)
+
+
+### Bug Fixes
+
+* **storage:** use POSIX separators for artifact keys ([1e5fa6f](https://github.com/ducktors/turborepo-remote-cache/commit/1e5fa6f3dc4427a33a7935d0c9dcffda09032b29)), closes [#800](https://github.com/ducktors/turborepo-remote-cache/issues/800)
+
 ## [2.11.3](https://github.com/ducktors/turborepo-remote-cache/compare/v2.11.2...v2.11.3) (2026-07-21)